Remove TU/e from LICENSE.md
The TU/e's copyright rests on the Python code in the quickstart for the EURO meets NeurIPS competition. We have never relied on that code in the PyVRP repository, so there is no TU/e copyright present in this repository. As such, the copyright attribution was not appropriate and this commit removes it.